CurrentTransactionsPage was moved from 'homepage' to 'budgetpage'
